+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+Fixes:
+- stop mds3 init retry for any validation step error that is non-recoverable
+
 
 ## 2.191.3